% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishdata.ices.R
\name{fishdata.ices}
\alias{fishdata.ices}
\title{Get limits values for  stocks provided by ICES stocksmart packages}
\usage{
fishdata.ices(
  Stock_Name = NULL,
  Assessment_Year = NULL,
  update = FALSE,
  from = 2017,
  to = NULL
)
}
\arguments{
\item{Stock_Name}{to get data for a specified stock}

\item{Assessment_Year}{to get data for a specified stock}
}
\description{
Get limits values for  stocks provided by ICES stocksmart packages
}
\examples{
limits<-limits.ices(Stock_Name='Albacore - North Atlantic',Assessment_Year=2020)
}
