% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishdata.noaa.R
\name{fishdata.noaa}
\alias{fishdata.noaa}
\title{Get fishadata values for stocks provided by NOAA stocksmart packages}
\usage{
fishdata.noaa(Stock_Name = NULL, Assessment_Year = NULL, update = FALSE)
}
\arguments{
\item{Stock_Name}{to get data for a specified stock}

\item{Assessment_Year}{to get data for a specified stock}
}
\description{
Get fishadata values for stocks provided by NOAA stocksmart packages
}
\examples{
limits<-fishdata.noaa(Stock_Name='Albacore - North Atlantic',Assessment_Year=2020)

}
