% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishdata.other.R
\name{fishdata.other}
\alias{fishdata.other}
\title{Get fishadata values for stocks provided by NOAA stocksmart packages}
\usage{
fishdata.other(
  Stock_Name = NULL,
  Assessment_Year = NULL,
  update = FALSE,
  user = NULL,
  password = NULL,
  server = NULL,
  db = NULL
)
}
\arguments{
\item{Stock_Name}{to get data for a specified stock}

\item{Assessment_Year}{to get data for a specified stock}
}
\description{
Get fishadata values for stocks provided by NOAA stocksmart packages
}
\examples{
fishdata.noaa.dta<-fishdata.noaa(Stock_Name='Albacore - North Atlantic',Assessment_Year=2020)

}
