% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area.R
\name{get_area}
\alias{get_area}
\title{Get sf object of all stocks area available}
\usage{
get_area(
  Stock_Name = NULL,
  update = FALSE,
  user = NULL,
  password = NULL,
  server = NULL,
  db = NULL
)
}
\arguments{
\item{Stock_Name}{to get data for a specified stock}

\item{Assessment_Year}{to get data for a specified stock}
}
\description{
Get sf object of all stocks area available
}
