% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.ices.R
\name{limits.ices}
\alias{limits.ices}
\title{Get limits values for  stocks provided by ICES stocksmart packages}
\usage{
limits.ices(
  Stock_Name = NULL,
  Assessment_Year = NULL,
  update = FALSE,
  from = 2018,
  to = NULL,
  average.nbyear = NULL,
  exclude = NULL
)
}
\arguments{
\item{Stock_Name}{to get data for a specified stock}

\item{Assessment_Year}{to get data for a specified stock}
}
\description{
Get limits values for  stocks provided by ICES stocksmart packages
}
\examples{
limits<-limits.ices(Stock_Name='fle.27.3a4',from=2025) from=2024
}
