% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.transcript.R
\name{score.transcript}
\alias{score.transcript}
\title{Get limits values for  stocks provided by ICES stocksmart packages}
\usage{
score.transcript(system.1.2.dta = NULL)
}
\arguments{
\item{system.1.2}{date provided by system.1.2 function}
}
\description{
Get limits values for  stocks provided by ICES stocksmart packages
}
\examples{
score<-score.transcript(system.2(ssci_name='SQUATINA SQUATINA',area='27.'))
}
