% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shi.R
\name{shi}
\alias{shi}
\title{Get SHI values by fleets using either totvallandg (values in euros) or totwghtlandg (landings weight)}
\usage{
shi(
  fleets.aer = NULL,
  stockdef = NULL,
  limits = NULL,
  fishdata = NULL,
  limit.assYear = NULL
)
}
\arguments{
\item{fleets.aer}{a set of aer data}

\item{stockdef}{stock definition DF}

\item{limits}{limits for stocks}

\item{fishdata}{time series for stocks}
}
\description{
Get SHI values by fleets using either totvallandg (values in euros) or totwghtlandg (landings weight)
}
\examples{
library(stocksmart)
library(dplyr)
library(sf)
library(ggplot2)
library(tidyr)#' test <-shi(fleet.aer.samp,NULL,NULL,NULL,2024)
ggplot(test) + geom_bar(aes(x=year,y=shi,fill=variable_code),stat='identity',position='dodge')

}
