% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockdef.noaa.R
\name{stockdef.noaa}
\alias{stockdef.noaa}
\title{Get limits values for  stocks provided by NOAA stocksmart packages}
\usage{
stockdef.noaa(Stock_Name = NULL, Assessment_Year = NULL, update = FALSE)
}
\arguments{
\item{Stock_Name}{to get data for a specified stock}

\item{Assessment_Year}{to get data for a specified stock}
}
\description{
Get limits values for  stocks provided by NOAA stocksmart packages
}
\examples{
limits<-limits.noaa(Stock_Name='Albacore - North Atlantic',Assessment_Year=2020)
}
