% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockdef.other.R
\name{stockdef.other}
\alias{stockdef.other}
\title{Get limits values for  stocks provided by ICES stocksmart packages}
\usage{
stockdef.other(
  Stock_Name = NULL,
  Assessment_Year = NULL,
  update = FALSE,
  user = NULL,
  password = NULL,
  server = NULL,
  db = NULL
)
}
\arguments{
\item{Stock_Name}{to get data for a specified stock}

\item{Assessment_Year}{to get data for a specified stock}
}
\description{
Get limits values for  stocks provided by ICES stocksmart packages
}
\examples{
limits<-limits.ices(Stock_Name='Albacore - North Atlantic',Assessment_Year=2020)
}
