% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.1.2.R
\name{system.1.2}
\alias{system.1.2}
\title{Get limits values for  stocks provided by ICES stocksmart packages}
\usage{
system.1.2(
  sci_name = NULL,
  area = NULL,
  stockdef = NULL,
  limits = NULL,
  fishdata = NULL,
  iucn.dta = NULL,
  sensitive.dta = NULL,
  area.dta = NULL,
  iucn_to_stock_area = NULL,
  iucn.token = NULL,
  nbyear_average = 6
)
}
\arguments{
\item{area}{catches area mentionned for the product}

\item{scientific_name}{taxonomic name of the product}
}
\description{
Get limits values for  stocks provided by ICES stocksmart packages
}
\examples{
all<-system.1.2(sci_name='SQUATINA SQUATINA',area='27.')
}
