% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.1.iucnws.R
\name{system.1.iucnws}
\alias{system.1.iucnws}
\title{Get limits values for  stocks provided by IUCN API through Rpackage rredlist}
\usage{
system.1.iucnws(
  sci_name = NULL,
  area = NULL,
  iucn.dta = NULL,
  sensitive.dta = NULL,
  area.dta = NULL,
  iucn_to_stock_area = NULL,
  token = NULL,
  longitude = NULL,
  latitude = NULL
)
}
\arguments{
\item{area}{catches area mentionned for the product}

\item{scientific_name}{taxonomic name of the product}
}
\description{
Get limits values for  stocks provided by IUCN API through Rpackage rredlist
}
\examples{
pre.score.2<-system.2(sci_name='SQUATINA SQUATINA',area='27.')
}
