% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.1.R
\name{system.1}
\alias{system.1}
\title{Get limits values for  stocks provided by ICES stocksmart packages}
\usage{
system.1(
  sci_name = NULL,
  area = NULL,
  iucn.dta = NULL,
  sensitive.dta = NULL,
  area.dta = NULL,
  iucn_to_stock_area = NULL,
  longitude = NULL,
  latitude = NULL
)
}
\arguments{
\item{area}{catches area mentionned for the product}

\item{scientific_name}{taxonomic name of the product}
}
\description{
Get limits values for  stocks provided by ICES stocksmart packages
}
\examples{
pre.score.2<-system.(sci_name='LOLIGO VULGARIS',area='31.1.2')
}
