% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.2.R
\name{system.2}
\alias{system.2}
\title{Get limits values for  stocks provided by ICES stocksmart packages}
\usage{
system.2(
  sci_name = NULL,
  area = NULL,
  stockdef = NULL,
  limits = NULL,
  fishdata = NULL,
  nbyear_average = 6,
  fmsy_upper = FALSE
)
}
\arguments{
\item{area}{catches area mentionned for the product}

\item{scientific_name}{taxonomic name of the product}
}
\description{
Get limits values for  stocks provided by ICES stocksmart packages
}
\examples{
pre.score.2<-system.2(sci_name='MOLVA MOLVA',area='27.1')
}
