% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom.impact.R
\name{bottom.impact}
\alias{bottom.impact}
\title{Get limits values for  stocks provided by ICES stocksmart packages}
\usage{
bottom.impact(gear.req = NULL, species.req = NULL)
}
\arguments{
\item{gear.req}{FAO code of gear}

\item{species.req}{Alpha code of species}
}
\description{
Get limits values for  stocks provided by ICES stocksmart packages
}
\examples{
impact<-bottom.impact(gear.req='PS',species.req='ANE')
}
