% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.transcript.R
\name{score.transcript}
\alias{score.transcript}
\title{Get limits values for  stocks provided by ICES stocksmart packages}
\usage{
\method{score}{transcript}(system.1.2.dta = NULL)
}
\arguments{
\item{system.1.2.dta}{data provided by system.1.2 function}
}
\description{
Get limits values for  stocks provided by ICES stocksmart packages
}
\examples{
score<-score.transcript(system.2(sci_name='SQUATINA SQUATINA',area='27.'))
}
